Refactor story imports
