+++ packages/react/checkbox/src/Checkbox.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css, keyframes } from 'testing/stitches.config';
 import { Label as LabelPrimitive } from '@radix-ui/react-label';
-import { RECOMMENDED_CSS__LABEL__ROOT } from '../../label/src/Label.stories';
+import { RECOMMENDED_CSS__LABEL__ROOT } from 'testing/label';
 import * as Checkbox from '@radix-ui/react-checkbox';
 
 export default { title: 'Components/Checkbox' };
