+++ packages/react/context-menu/src/ContextMenu.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { css, keyframes } from 'testing/stitches.config';
 import * as ContextMenu from '@radix-ui/react-context-menu';
 import { foodGroups } from 'testing/foods';
-import { classes, TickIcon } from '../../menu/src/Menu.stories';
+import { classes, TickIcon } from 'testing/menu';
 
 const { contentClass, itemClass, labelClass, separatorClass, subTriggerClass } = classes;
