+++ packages/react/radio-group/src/RadioGroup.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css, keyframes } from 'testing/stitches.config';
 import { Label as LabelPrimitive } from '@radix-ui/react-label';
-import { RECOMMENDED_CSS__LABEL__ROOT } from '../../label/src/Label.stories';
+import { RECOMMENDED_CSS__LABEL__ROOT } from 'testing/label';
 import { DirectionProvider } from '@radix-ui/react-direction';
 import * as RadioGroup from '@radix-ui/react-radio-group';
