+++ packages/react/switch/src/Switch.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { css } from 'testing/stitches.config';
 import { Label as LabelPrimitive } from '@radix-ui/react-label';
-import { RECOMMENDED_CSS__LABEL__ROOT } from '../../label/src/Label.stories';
+import { RECOMMENDED_CSS__LABEL__ROOT } from 'testing/label';
 import * as Switch from '@radix-ui/react-switch';
 
 export default { title: 'Components/Switch' };
