+++ testing/src/label.ts
@@ -1,0 +1,8 @@
+export const RECOMMENDED_CSS__LABEL__ROOT = {
+  // ensures it can receive vertical margins
+  display: 'inline-block',
+  // better default alignment
+  verticalAlign: 'middle',
+  // mimics default `label` tag (as we render a `span`)
+  cursor: 'default',
+};
